backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pandas==2.1.4
chromadb==0.5.0
sentence-transformers==2.2.2
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
ollama==0.1.7
google-generativeai==0.3.2
numpy==1.24.3
scikit-learn==1.3.2
tokenizers==0.13.3
transformers==4.30.0
crewai
litellm
